Android 17.0.0 release 1

# -----BEGIN PGP SIGNATURE-----
#
# iF0EABECAB0WIQRDQNE1cO+UXoOBCWTorT+BmrEOeAUCajHazQAKCRDorT+BmrEO
# eOdtAJ9/ibQg4cAUN5+nCzA0MJ1C/A1OxQCcDa+omPbfz6q5eJvuopmxTC7YbPw=
# =AFaZ
# -----END PGP SIGNATURE-----
# gpg: Signature made Wed 17 Jun 2026 02:22:53 IDT
# gpg:                using DSA key 4340D13570EF945E83810964E8AD3F819AB10E78
# gpg: Good signature from "The Android Open Source Project <initial-contribution@android.com>" [unknown]
# gpg: WARNING: This key is not certified with a trusted signature!
# gpg:          There is no indication that the signature belongs to the owner.
# Primary key fingerprint: 4340 D135 70EF 945E 8381  0964 E8AD 3F81 9AB1 0E78

* tag 'android-17.0.0_r1': (2173 commits)
  [graphite] WrapTexture does not require SkColorType, can force opaque for any color type
  [graphite] Use TextureFormatXferFn instead of Caps::supportedTransferColorType()
  [graphite] Remove UploadSource from Texture::canUploadOnHost
  [graphite] Add TextureProxy::format() helper function
  Roll SKP CIPD package from 557 to 558
  Manual roll ANGLE from 0755fa45413a to 8d8796935e4f (11 revisions)
  Roll recipe dependencies (trivial).
  [graphite] Plumb TextureProxyView through read/upload codepaths
  Use a local data copy for strike deserialization
  Terracotta-Phase-1: Reorder member destruction
  Add a cache of font variation typefaces to SkParagraph's FontCollection
  Update fiddler-base for 93dc3a99a732
  Avoid duplicate calls to the same logic code.
  Use SkSafeMath to prevent overflow in pixel offset calculations.
  Manual roll Dawn from 012f6e92c62b to 0d2b886d015d (7 revisions)
  Revert "Improve readibility of SkCpu and add documentation"
  Roll recipe dependencies (trivial).
  Roll ANGLE from 95a343232b1b to 0755fa45413a (9 revisions)
  Roll Skia Infra from b555f647994c to 08b8f9693e62 (17 revisions)
  Roll Dawn from 4cc0cdeb1d3b to 012f6e92c62b (10 revisions)
  ...

+++ bench/TypefaceBench.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Google Inc.
+ * Copyright 2018 Google LLC
  *
  * Use of this source code is governed by a BSD-style license that can be
  * found in the LICENSE file.
@@ -245,7 +245,7 @@ protected:
         for (int i = 0; i < loops * 3; ++i) {
             for (auto& line : fLines) {
                 font.textToGlyphs(line->utf.data(), line->utf.size(), fEncoding,
-                                  {fGlyphIds.data(), line->glyphCount});
+                                  {fGlyphIds.data(), (size_t)line->glyphCount});
             }
         }
     }
