+++ gm/scaledemoji.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Google Inc.
+ * Copyright 2018 Google LLC
  *
  * Use of this source code is governed by a BSD-style license that can be
  * found in the LICENSE file.
@@ -29,7 +29,7 @@ static sk_sp<SkTextBlob> make_hpos_test_blob_utf8(const char* text, const SkFont
     constexpr SkTextEncoding enc = SkTextEncoding::kUTF8;
     SkTextBlobBuilder builder;
     size_t len = strlen(text);
-    int glyphCount = font.countText(text, len, enc);
+    size_t glyphCount = font.countText(text, len, enc);
     const auto& buffer = builder.allocRunPosH(font, glyphCount, 0);
     (void)font.textToGlyphs(text, len, enc, {buffer.glyphs, glyphCount});
     font.getXPos({buffer.glyphs, glyphCount}, {buffer.pos, glyphCount});
